chore: Update npm dependencies
This commit updates the npm dependencies in the package.json file. It includes updates to the following packages:
- @firebase/app: from version 0.10.7 to version 1.7.5
- @firebase/auth: from version 1.7.5 to version 1.7.5
- firebase: from version 10.12.4 to version 10.12.4
- framer-motion: from version 11.3.12 to version 11.3.12
- react-router-dom: from version 6.25.1 to version 6.25.1
- react-scripts: from version 3.0.1 to version 3.0.1
- react-toastify: from version 10.0.5 to version 10.0.5

This change ensures that the project is using the latest stable versions of the dependencies, improving compatibility and taking advantage of any bug fixes or new features.

Refactor App.js to use GlobalProvider

This commit updates the App.js file to import and use the GlobalProvider component from GlobalContext.js. The GlobalProvider component is wrapped around the entire application, providing access to the global state variables for the user's watchlist. This change ensures that the watchlist data is available to all components that need it.

Fix bug with toast notification in SearchResultsList

This commit fixes a bug in the SearchResultsList component where the toast notification for adding a film to the watchlist was not working correctly. The toast.success function call has been removed from the code, simplifying the code and removing unnecessary visual feedback for the user.

Refactor date formatting in SingleMoovies and SingleSeries components

This commit refactors the date formatting logic in the SingleMoovies and SingleSeries components. It adds a check to handle cases where the date string is empty, preventing errors in the formatting function. This change improves the reliability and stability of the components.

Fix watchlist success popup

This commit fixes an issue with the watchlist success popup, ensuring that it is displayed correctly when a film is added to the watchlist.

Update local storage key for watchlist to include user ID and fix problem with detail page

This commit updates the local storage key for the watchlist to include the user ID. The user ID is obtained from the authentication module and used to create a unique key for each user's watchlist. This change ensures that each user's watchlist is stored separately and prevents conflicts between different users' watchlists. Additionally, a problem with the detail page has been fixed.

Add pagination component to search component

This commit adds a pagination component to the search component, allowing users to navigate through multiple pages of search results.