chore: Update firebase.json to use "build" folder for hosting
This commit updates the firebase.json file to change the "public" property from "public" to "build". This change ensures that the correct folder is used for hosting the application. The "build" folder typically contains the optimized and production-ready version of the application, so using it for hosting improves page load performance and ensures that the latest changes are deployed to the hosting environment.